feat: Reading datas from tcp connections
when data it's sent over a network, it's sent in packets. Each message is split into packets, the packent are sent, they arrive out of order (pottentialy), and they are reassembled on the other side. And without a protocol like TCP, you can't guarantee that the order is correct. Tcp allows orderer data to be safaely sent across the internet